adaugare modificari pentru hibernate

+++ internship-rest/src/main/java/com/endava/internship/rest/service/BuildingService.java
@@ -1,11 +1,11 @@
 package com.endava.internship.rest.service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.endava.internship.rest.exception.MyException;
 import com.endava.internship.rest.exception.ValidationException;
 import com.endava.internship.rest.model.Building;
 import com.endava.internship.rest.repository.BuildingRepository;
@@ -31,7 +31,14 @@ public class BuildingService
 	
 	public void deleteById(Integer id)
 	{
-		buildingRepository.deleteById(id);	
+		if(buildingRepository.getById(id)!= null)
+		{
+			buildingRepository.deleteById(id);	
+		}
+		else
+		{
+			throw new MyException("Building doesn't exists!");
+		}
 	}
 	
 	public void deleteAll()
@@ -47,6 +54,12 @@ public class BuildingService
 		}
 		Building build = this.buildingRepository.createBuilding(building);
 		return build;
+	}
+	
+	public Building updateBuilding(Building building, Integer buildingId)
+	{
+		Building newBuilding = this.buildingRepository.updateBuilding(building, buildingId);
+		return newBuilding;
 	}
 	
 	public List<Building> createBuildings(List<Building> list)
